requirements.txt
asgiref==3.6.0
charset-normalizer==3.1.0
coverage==7.2.6
Cython==0.29.34
Django==4.2.1
django-admin-sortable2==2.1.8
django-grappelli==3.0.6
docutils==0.20.1
grappelli-nested-inlines==0.6.1
psycopg2-binary==2.9.6
pypugjs==5.9.12
python-dotenv==1.0.0
six==1.16.0
sqlparse==0.4.4
